Refactored model declaration code
